+++ src/main/java/by/fiveelement/api/login/ForgotPasswordApiRequest.java
@@ -1,4 +1,4 @@
-package by.fiveelement;
+package by.fiveelement.api.login;
 
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
